Update recaptcha.adoc
Google has changed some things, along with browsers are now a bit stricter and require a new 'unsafe-inline' setting.   With the values in the current screenshot, the following browsers error with:

* Chrome: `Refused to load the script 'https://www.gstatic.com...' because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-inline' ...`
* Firefox: `Content Security Policy: The page’s settings blocked the loading of a resource at inline (“script-src”).`

Also noting that it should be a v2 "checkbox" API key. 

I'm unable to update the screenshot you have since my data is different than the demo data in the screenshot.